Not sure what I want to focus on
fixing special moves?
reading warriors from a file?
implementing all warriors / specials / skills?
React?
